[FIX] purchase_receipt_expectation: Use odoo_test_helper
